商家查看订单相关业务

+++ src/main/java/sspu/informationsystem/entity/Dishes.java
@@ -10,7 +10,7 @@ public class Dishes implements Serializable {
 
     private String dName;
 
-    private BigDecimal dPrice;
+    private Double dPrice;
 
     private String dDetails;
